requirements.txt

numpy==1.16.0
pandas==0.23.4
pytest==4.2.0
python-dateutil==2.7.3
biopython==1.72
matplotlib==2.2.3
